only delete abandoned cart when status is not pending

+++ includes/processes/class-mailchimp-woocommerce-single-order.php
@@ -142,8 +142,10 @@ class MailChimp_WooCommerce_Single_Order extends Mailchimp_Woocommerce_Job
                 }
             }
 
-            // delete the AC cart record.
-            $deleted_abandoned_cart = !empty($this->cart_session_id) && $api->deleteCartByID($store_id, $this->cart_session_id);
+            if ($order->getOriginalWooStatus() !== 'pending') {
+                // delete the AC cart record.
+                $deleted_abandoned_cart = !empty($this->cart_session_id) && $api->deleteCartByID($store_id, $this->cart_session_id);
+            }
 
             // skip amazon orders and skip privacy protected orders.
             if ($order->isFlaggedAsAmazonOrder()) {
